修改预订单dto：敏感数据由后端远程调用获取

+++ order-service/src/main/java/org/rail/orderservice/model/entity/PreOrderDetails.java
@@ -14,7 +14,7 @@ public class PreOrderDetails {
     private Long preOrderId;
     private String realName;
     // 证件类型：0-身份证 1-护照 2-港澳通行证
-    private Long idType;
+    private Integer idType;
     private String idCard;
     // 车票类型：0-成人票 1-儿童票 2-学生票 3-残疾军人
     private Integer ticketType;
